public_interface/pyproject.toml
[project]
name = "autonomous-email-inbox"
version = "1.0.0"
description = "Autonomous Email Inbox with LangSmith Integration"
requires-python = ">=3.9"
dependencies = [
    "Flask>=2.0.0,<3.0.0",
    "requests>=2.25.0",
    "python-dotenv>=0.19.0",
    "gunicorn>=20.0.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["public_interface"]
